Added support for debugMessage modes that determine how debug messages are outputted.
A new config.yml is generated with two settings: debugMessageMode and debugTargets.

debugMessageMode determines how debug messages are outputted. Possible settings are 'all' (output to all players on the server), 'first' (output to the first player on the server), 'targets' (output to a list of targets) and 'console' (the default, output to the console).

debugTargets is simply a comma separated list of targets that are only used if debugMessageMode is set to 'targets'.